requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Werkzeug==2.1.0
openai==0.27.0
google-cloud-vision==2.6.0
